requirements.txt
azure-cognitiveservices-speech>=1.25.0
